style: unify websocket implementations
